+++ src/js/components/sidebar/FeaturesSet.js
@@ -1,4 +1,3 @@
-'use strict';
 import React from 'react';
 import ReactDOM from 'react-dom';
