+++ src/js/components/sidebar/Slider.js
@@ -1,4 +1,3 @@
-'use strict';
 import React from 'react';
 
 class Slider extends React.Component {
