+++ src/js/components/table/SearchResult.js
