+++ src/js/components/table/TableTabs.js
